+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Don't report skipped tasks as YAML parser error.
+
 ## [0.14.2] - 2020-09-23
 
 ### Fixed